// app/types.ts
export interface AppInfo {
  email: string;
  password: string;
  passwordConfirm: string;
  phone: string;

  accessToken: string;
  refreshToken: string;

  isAdmin: boolean;

  language: string;

  userType: string;
}

export interface BaseInfo {
  width: number;
  height: number;
}